Parte de servicios interfaces, mucha pereza, mañana esta acabado...

+++ app/src/main/java/com/example/pmdm_practicasexamen/EJ1/models/Info.java
@@ -7,7 +7,7 @@ public class Info {
     @SerializedName("count")
     private long count;
     @SerializedName("pages")
-    private long pages;
+    private int pages;
     @SerializedName("next")
     private String next;
     @SerializedName("prev")
@@ -16,8 +16,8 @@ public class Info {
     public long getCount() { return count; }
     public void setCount(long value) { this.count = value; }
 
-    public long getPages() { return pages; }
-    public void setPages(long value) { this.pages = value; }
+    public int getPages() { return pages; }
+    public void setPages(int value) { this.pages = value; }
 
     public String getNext() { return next; }
     public void setNext(String value) { this.next = value; }
